clippy

+++ src/aoc2022/day10.rs
@@ -157,7 +157,7 @@ impl CycleChecker {
 
             let signal_strength = register.value;
             self.signal_str_at
-                .push(cycle as isize * signal_strength as isize);
+                .push(cycle as isize * signal_strength);
         }
     }
 }
